4.15 to 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,13 +1,7 @@
 const blogRouter = require('express').Router()
-const { request } = require('express')
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
-const getRandomUser = async () => {
-    const count = await User.countDocuments()
-    const random = Math.floor(Math.random() * count)
-    return User.findOne().skip(random)
-}
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -21,9 +15,8 @@ blogRouter.get('/:id', async (request, response) => {
 
 blogRouter.post('/', async (request, response) => {
     const body = request.body
-    const randomUser = await getRandomUser()
 
-    if (!randomUser) {
+    if (!request.user) {
         return response.status(400).json({ error: 'userId missing or not valid' })
     }
     if (!request.body.title || !request.body.url) {
@@ -35,18 +28,25 @@ blogRouter.post('/', async (request, response) => {
         "author": body.author,
         "url": body.url,
         "likes": body.likes,
-        "user": randomUser._id
+        "user": request.user._id
     })
-    
+
     const savedBlog = await blog.save()
-    randomUser.blogs = randomUser.blogs.concat(savedBlog._id)
-    await randomUser.save()
+    request.user.blogs = request.user.blogs.concat(savedBlog._id)
+    await request.user.save()
     response.status(201).json(savedBlog)
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    if (request.user.id.toString() === blogToDelete.user.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    }
+    else {
+        response.status(403).end()
+    }
 })
 
 blogRouter.put('/:id', async (request, response) => {
